requirements.txt
requests
colorama
nimporter
pyyaml
tldextract
fuzzywuzzy
